Major changes:
It's been a long time since we last commited a change. That's because we have been trying a lot of ideas until finally one made sense. This major change comes with a new approach.

exp2_bothhospital_indicator: a new experiment where we try what happens if the data is labeled and we know from which hospital is coming from.

exp2_transferlearning: a new experiment trying to learn from only one hospital and predicting the other one.

extract_peaks: auxiliary function used to extract interesting peaks into an excel form.

hrc_model: a new script that does the same as gm model but with HRC data.