PUBLIC DATASETS:
We made public the datasets!

The pkls have the data already preprocessed to reproduce the paper results including training and testing folds. The raw dataset will be uploaded in the future in a zip format.

We also have updated the scrips that do the preprocess to the raw dataset.

We also removed the R script which was used to translate from Bruker to MZML because we are going to upload directly the MZML files.